% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Address}
\alias{Address}
\title{Address Object}
\usage{
Address(metadata = NULL, formattedValue = NULL, type = NULL,
  formattedType = NULL, poBox = NULL, streetAddress = NULL,
  extendedAddress = NULL, city = NULL, region = NULL, postalCode = NULL,
  country = NULL, countryCode = NULL)
}
\arguments{
\item{metadata}{Metadata about the address}

\item{formattedValue}{The read-only value of the address formatted in the viewer's account locale or the `Accept-Language` HTTP header locale}

\item{type}{The type of the address}

\item{formattedType}{The read-only type of the address translated and formatted in the viewer's account locale or the `Accept-Language` HTTP header locale}

\item{poBox}{The P}

\item{streetAddress}{The street address}

\item{extendedAddress}{The extended address of the address; for example, the apartment number}

\item{city}{The city of the address}

\item{region}{The region of the address; for example, the state or province}

\item{postalCode}{The postal code of the address}

\item{country}{The country of the address}

\item{countryCode}{The [ISO 3166-1 alpha-2](http://www}
}
\value{
Address object
}
\description{
Address Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's physical address. May be a P.O. box or street address. All fields are optional.
}
\seealso{
Other res: \code{\link{Biography}}, \code{\link{Birthday}},
  \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

