% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{AgeRangeType}
\alias{AgeRangeType}
\title{AgeRangeType Object}
\usage{
AgeRangeType(metadata = NULL, ageRange = NULL)
}
\arguments{
\item{metadata}{Metadata about the age range}

\item{ageRange}{The age range}
}
\value{
AgeRangeType object
}
\description{
AgeRangeType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's age range.
}
