% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Biography}
\alias{Biography}
\title{Biography Object}
\usage{
Biography(contentType = NULL, metadata = NULL, value = NULL)
}
\arguments{
\item{contentType}{The content type of the biography}

\item{metadata}{Metadata about the biography}

\item{value}{The short biography}
}
\value{
Biography object
}
\description{
Biography Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's short biography.
}
