% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Birthday}
\alias{Birthday}
\title{Birthday Object}
\usage{
Birthday(metadata = NULL, text = NULL, date = NULL)
}
\arguments{
\item{metadata}{Metadata about the birthday}

\item{text}{A free-form string representing the user's birthday}

\item{date}{The date of the birthday}
}
\value{
Birthday object
}
\description{
Birthday Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's birthday. At least one of the `date` and `text` fields arespecified. The `date` and `text` fields typically represent the samedate, but are not guaranteed to.
}
