% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{ContactGroupMembership}
\alias{ContactGroupMembership}
\title{ContactGroupMembership Object}
\usage{
ContactGroupMembership(contactGroupId = NULL)
}
\arguments{
\item{contactGroupId}{The contact group ID for the contact group membership}
}
\value{
ContactGroupMembership object
}
\description{
ContactGroupMembership Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google contact group membership.
}
