% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{CoverPhoto}
\alias{CoverPhoto}
\title{CoverPhoto Object}
\usage{
CoverPhoto(metadata = NULL, default = NULL, url = NULL)
}
\arguments{
\item{metadata}{Metadata about the cover photo}

\item{default}{True if the cover photo is the default cover photo;}

\item{url}{The URL of the cover photo}
}
\value{
CoverPhoto object
}
\description{
CoverPhoto Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's read-only cover photo. A large image shown on the person'sprofile page that represents who they are or what they care about.
}
