% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Date}
\alias{Date}
\title{Date Object}
\usage{
Date(year = NULL, day = NULL, month = NULL)
}
\arguments{
\item{year}{Year of date}

\item{day}{Day of month}

\item{month}{Month of year}
}
\value{
Date object
}
\description{
Date Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a whole calendar date, for example a date of birth. The timeof day and time zone are either specified elsewhere or are notsignificant. The date is relative to the[Proleptic Gregorian Calendar](https://en.wikipedia.org/wiki/Proleptic_Gregorian_calendar).The day may be 0 to represent a year and month where the day is notsignificant. The year may be 0 to represent a month and day independentof year; for example, anniversary date.
}
