% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{EmailAddress}
\alias{EmailAddress}
\title{EmailAddress Object}
\usage{
EmailAddress(displayName = NULL, metadata = NULL, type = NULL,
  value = NULL, formattedType = NULL)
}
\arguments{
\item{displayName}{The display name of the email}

\item{metadata}{Metadata about the email address}

\item{type}{The type of the email address}

\item{value}{The email address}

\item{formattedType}{The read-only type of the email address translated and formatted in the}
}
\value{
EmailAddress object
}
\description{
EmailAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's email address.
}
