% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Event}
\alias{Event}
\title{Event Object}
\usage{
Event(metadata = NULL, type = NULL, date = NULL, formattedType = NULL)
}
\arguments{
\item{metadata}{Metadata about the event}

\item{type}{The type of the event}

\item{date}{The date of the event}

\item{formattedType}{The read-only type of the event translated and formatted in the}
}
\value{
Event object
}
\description{
Event Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event related to the person.
}
