% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{FieldMetadata}
\alias{FieldMetadata}
\title{FieldMetadata Object}
\usage{
FieldMetadata(source = NULL, verified = NULL, primary = NULL)
}
\arguments{
\item{source}{The source of the field}

\item{verified}{True if the field is verified; false if the field is unverified}

\item{primary}{True if the field is the primary field; false if the field is a secondary}
}
\value{
FieldMetadata object
}
\description{
FieldMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about a field.
}
