% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_functions.R
\name{getBatchGet}
\alias{getBatchGet}
\title{Provides information about a list of specific people by specifying a list of requested resource names. Use `people/me` to indicate the authenticated user.}
\usage{
getBatchGet(resourceNames = NULL, requestMask.includeField = NULL)
}
\arguments{
\item{resourceNames}{The resource name, such as one returned by [`people}

\item{requestMask.includeField}{Comma-separated list of fields to be included in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/contacts
\item https://www.googleapis.com/auth/contacts.readonly
\item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/user.addresses.read
\item https://www.googleapis.com/auth/user.birthday.read
\item https://www.googleapis.com/auth/user.emails.read
\item https://www.googleapis.com/auth/user.phonenumbers.read
\item https://www.googleapis.com/auth/userinfo.email
\item https://www.googleapis.com/auth/userinfo.profile
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/contacts, https://www.googleapis.com/auth/contacts.readonly, https://www.googleapis.com/auth/plus.login, https://www.googleapis.com/auth/user.addresses.read, https://www.googleapis.com/auth/user.birthday.read, https://www.googleapis.com/auth/user.emails.read, https://www.googleapis.com/auth/user.phonenumbers.read, https://www.googleapis.com/auth/userinfo.email, https://www.googleapis.com/auth/userinfo.profile)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/people/}{Google Documentation}
}

