% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{GetPeopleResponse}
\alias{GetPeopleResponse}
\title{GetPeopleResponse Object}
\usage{
GetPeopleResponse(responses = NULL)
}
\arguments{
\item{responses}{The response for each requested resource name}
}
\value{
GetPeopleResponse object
}
\description{
GetPeopleResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
