% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{ImClient}
\alias{ImClient}
\title{ImClient Object}
\usage{
ImClient(formattedProtocol = NULL, formattedType = NULL, metadata = NULL,
  type = NULL, protocol = NULL, username = NULL)
}
\arguments{
\item{formattedProtocol}{The read-only protocol of the IM client formatted in the viewer's account}

\item{formattedType}{The read-only type of the IM client translated and formatted in the}

\item{metadata}{Metadata about the IM client}

\item{type}{The type of the IM client}

\item{protocol}{The protocol of the IM client}

\item{username}{The user name used in the IM client}
}
\value{
ImClient object
}
\description{
ImClient Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's instant messaging client.
}
