% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Interest}
\alias{Interest}
\title{Interest Object}
\usage{
Interest(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the interest}

\item{value}{The interest; for example, `stargazing`}
}
\value{
Interest object
}
\description{
Interest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One of the person's interests.
}
