% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{ListConnectionsResponse}
\alias{ListConnectionsResponse}
\title{ListConnectionsResponse Object}
\usage{
ListConnectionsResponse(nextPageToken = NULL, connections = NULL,
  nextSyncToken = NULL)
}
\arguments{
\item{nextPageToken}{The token that can be used to retrieve the next page of results}

\item{connections}{The list of people that the requestor is connected to}

\item{nextSyncToken}{The token that can be used to retrieve changes since the last request}
}
\value{
ListConnectionsResponse object
}
\description{
ListConnectionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
