% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Nickname}
\alias{Nickname}
\title{Nickname Object}
\usage{
Nickname(metadata = NULL, type = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the nickname}

\item{type}{The type of the nickname}

\item{value}{The nickname}
}
\value{
Nickname object
}
\description{
Nickname Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's nickname.
}
