% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Occupation}
\alias{Occupation}
\title{Occupation Object}
\usage{
Occupation(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the occupation}

\item{value}{The occupation; for example, `carpenter`}
}
\value{
Occupation object
}
\description{
Occupation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's occupation.
}
