% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Organization}
\alias{Organization}
\title{Organization Object}
\usage{
Organization(endDate = NULL, symbol = NULL, name = NULL,
  metadata = NULL, location = NULL, title = NULL, current = NULL,
  formattedType = NULL, startDate = NULL, domain = NULL,
  department = NULL, phoneticName = NULL, type = NULL,
  jobDescription = NULL)
}
\arguments{
\item{endDate}{The end date when the person left the organization}

\item{symbol}{The symbol associated with the organization; for example, a stock ticker}

\item{name}{The name of the organization}

\item{metadata}{Metadata about the organization}

\item{location}{The location of the organization office the person works at}

\item{title}{The person's job title at the organization}

\item{current}{True if the organization is the person's current organization;}

\item{formattedType}{The read-only type of the organization translated and formatted in the}

\item{startDate}{The start date when the person joined the organization}

\item{domain}{The domain name associated with the organization; for example, `google}

\item{department}{The person's department at the organization}

\item{phoneticName}{The phonetic name of the organization}

\item{type}{The type of the organization}

\item{jobDescription}{The person's job description at the organization}
}
\value{
Organization object
}
\description{
Organization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's past or current organization. Overlapping date ranges arepermitted.
}
