% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Person}
\alias{Person}
\title{Person Object}
\usage{
Person(nicknames = NULL, names = NULL, relations = NULL,
  occupations = NULL, emailAddresses = NULL, organizations = NULL,
  etag = NULL, braggingRights = NULL, metadata = NULL,
  residences = NULL, genders = NULL, resourceName = NULL,
  interests = NULL, biographies = NULL, skills = NULL,
  relationshipStatuses = NULL, photos = NULL, ageRange = NULL,
  taglines = NULL, ageRanges = NULL, addresses = NULL, events = NULL,
  memberships = NULL, phoneNumbers = NULL, coverPhotos = NULL,
  imClients = NULL, birthdays = NULL, locales = NULL,
  relationshipInterests = NULL, urls = NULL)
}
\arguments{
\item{nicknames}{The person's nicknames}

\item{names}{The person's names}

\item{relations}{The person's relations}

\item{occupations}{The person's occupations}

\item{emailAddresses}{The person's email addresses}

\item{organizations}{The person's past or current organizations}

\item{etag}{The [HTTP entity tag](https://en}

\item{braggingRights}{The person's bragging rights}

\item{metadata}{Metadata about the person}

\item{residences}{The person's residences}

\item{genders}{The person's genders}

\item{resourceName}{The resource name for the person, assigned by the server}

\item{interests}{The person's interests}

\item{biographies}{The person's biographies}

\item{skills}{The person's skills}

\item{relationshipStatuses}{The person's relationship statuses}

\item{photos}{The person's photos}

\item{ageRange}{DEPRECATED(Please read person}

\item{taglines}{The person's taglines}

\item{ageRanges}{The person's age ranges}

\item{addresses}{The person's street addresses}

\item{events}{The person's events}

\item{memberships}{The person's group memberships}

\item{phoneNumbers}{The person's phone numbers}

\item{coverPhotos}{The person's cover photos}

\item{imClients}{The person's instant messaging clients}

\item{birthdays}{The person's birthdays}

\item{locales}{The person's locale preferences}

\item{relationshipInterests}{The kind of relationship the person is looking for}

\item{urls}{The person's associated URLs}
}
\value{
Person object
}
\description{
Person Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a person merged from various data sources such as theauthenticated user's contacts and profile data. Fields other than IDs,metadata, and group memberships are user-edited.Most fields can have multiple items. The items in a field have no guaranteedorder, but each non-empty field is guaranteed to have exactly one field with`metadata.primary` set to true.NEXT_ID: 31
}
