% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{PersonMetadata}
\alias{PersonMetadata}
\title{PersonMetadata Object}
\usage{
PersonMetadata(sources = NULL, previousResourceNames = NULL,
  deleted = NULL, objectType = NULL, linkedPeopleResourceNames = NULL)
}
\arguments{
\item{sources}{The sources of data for the person}

\item{previousResourceNames}{Any former resource names this person has had}

\item{deleted}{True if the person resource has been deleted}

\item{objectType}{DEPRECATED(Please read person}

\item{linkedPeopleResourceNames}{Resource names of people linked to this resource}
}
\value{
PersonMetadata object
}
\description{
PersonMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read-only metadata about a person.
}
