% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{PhoneNumber}
\alias{PhoneNumber}
\title{PhoneNumber Object}
\usage{
PhoneNumber(metadata = NULL, type = NULL, value = NULL,
  formattedType = NULL, canonicalForm = NULL)
}
\arguments{
\item{metadata}{Metadata about the phone number}

\item{type}{The type of the phone number}

\item{value}{The phone number}

\item{formattedType}{The read-only type of the phone number translated and formatted in the}

\item{canonicalForm}{The read-only canonicalized [ITU-T E}
}
\value{
PhoneNumber object
}
\description{
PhoneNumber Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's phone number.
}
