% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{ProfileMetadata}
\alias{ProfileMetadata}
\title{ProfileMetadata Object}
\usage{
ProfileMetadata(objectType = NULL)
}
\arguments{
\item{objectType}{The profile object type}
}
\value{
ProfileMetadata object
}
\description{
ProfileMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The read-only metadata about a profile.
}
