% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Relation}
\alias{Relation}
\title{Relation Object}
\usage{
Relation(metadata = NULL, person = NULL, type = NULL,
  formattedType = NULL)
}
\arguments{
\item{metadata}{Metadata about the relation}

\item{person}{The name of the other person this relation refers to}

\item{type}{The person's relation to the other person}

\item{formattedType}{The type of the relation translated and formatted in the viewer's account locale or the locale specified in the Accept-Language HTTP header}
}
\value{
Relation object
}
\description{
Relation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's relation to another person.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

