% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{RelationshipInterest}
\alias{RelationshipInterest}
\title{RelationshipInterest Object}
\usage{
RelationshipInterest(metadata = NULL, value = NULL, formattedValue = NULL)
}
\arguments{
\item{metadata}{Metadata about the relationship interest}

\item{value}{The kind of relationship the person is looking for}

\item{formattedValue}{The value of the relationship interest translated and formatted in the viewer's account locale or the locale specified in the Accept-Language HTTP header}
}
\value{
RelationshipInterest object
}
\description{
RelationshipInterest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The kind of relationship the person is looking for.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

