% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{RelationshipStatus}
\alias{RelationshipStatus}
\title{RelationshipStatus Object}
\usage{
RelationshipStatus(metadata = NULL, value = NULL, formattedValue = NULL)
}
\arguments{
\item{metadata}{Metadata about the relationship status}

\item{value}{The relationship status}

\item{formattedValue}{The read-only value of the relationship status translated and formatted in}
}
\value{
RelationshipStatus object
}
\description{
RelationshipStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's read-only relationship status.
}
