% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Residence}
\alias{Residence}
\title{Residence Object}
\usage{
Residence(metadata = NULL, value = NULL, current = NULL)
}
\arguments{
\item{metadata}{Metadata about the residence}

\item{value}{The address of the residence}

\item{current}{True if the residence is the person's current residence; false if the residence is a past residence}
}
\value{
Residence object
}
\description{
Residence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's past or current residence.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Skill}}, \code{\link{Source}},
  \code{\link{Tagline}}, \code{\link{Url}}
}

