% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Skill}
\alias{Skill}
\title{Skill Object}
\usage{
Skill(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the skill}

\item{value}{The skill; for example, `underwater basket weaving`}
}
\value{
Skill object
}
\description{
Skill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A skill that the person has.
}
