% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(profileMetadata = NULL, type = NULL, etag = NULL, id = NULL)
}
\arguments{
\item{profileMetadata}{Metadata about a source of type PROFILE}

\item{type}{The source type}

\item{etag}{The [HTTP entity tag](https://en}

\item{id}{The unique identifier within the source type generated by the server}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The source of a field.
}
