% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Tagline}
\alias{Tagline}
\title{Tagline Object}
\usage{
Tagline(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the tagline}

\item{value}{The tagline}
}
\value{
Tagline object
}
\description{
Tagline Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A read-only brief one-line description of the person.
}
