% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Url}
\alias{Url}
\title{Url Object}
\usage{
Url(formattedType = NULL, metadata = NULL, type = NULL, value = NULL)
}
\arguments{
\item{formattedType}{The read-only type of the URL translated and formatted in the viewer's}

\item{metadata}{Metadata about the URL}

\item{type}{The type of the URL}

\item{value}{The URL}
}
\value{
Url object
}
\description{
Url Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's associated URLs.
}
