% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Address}
\alias{Address}
\title{Address Object}
\usage{
Address(streetAddress = NULL, metadata = NULL, countryCode = NULL,
  formattedType = NULL, city = NULL, formattedValue = NULL,
  country = NULL, type = NULL, extendedAddress = NULL, poBox = NULL,
  postalCode = NULL, region = NULL)
}
\arguments{
\item{streetAddress}{The street address}

\item{metadata}{Metadata about the address}

\item{countryCode}{The [ISO 3166-1 alpha-2](http://www}

\item{formattedType}{The read-only type of the address translated and formatted in the viewer's}

\item{city}{The city of the address}

\item{formattedValue}{The unstructured value of the address}

\item{country}{The country of the address}

\item{type}{The type of the address}

\item{extendedAddress}{The extended address of the address; for example, the apartment number}

\item{poBox}{The P}

\item{postalCode}{The postal code of the address}

\item{region}{The region of the address; for example, the state or province}
}
\value{
Address object
}
\description{
Address Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's physical address. May be a P.O. box or street address. All fieldsare optional.
}
