% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{BraggingRights}
\alias{BraggingRights}
\title{BraggingRights Object}
\usage{
BraggingRights(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the bragging rights}

\item{value}{The bragging rights; for example, `climbed mount everest`}
}
\value{
BraggingRights object
}
\description{
BraggingRights Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's bragging rights.
}
