% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{EmailAddress}
\alias{EmailAddress}
\title{EmailAddress Object}
\usage{
EmailAddress(metadata = NULL, value = NULL, type = NULL,
  formattedType = NULL)
}
\arguments{
\item{metadata}{Metadata about the email address}

\item{value}{The email address}

\item{type}{The type of the email address}

\item{formattedType}{The read-only type of the email address translated and formatted in the viewer's account locale or the `Accept-Language` HTTP header locale}
}
\value{
EmailAddress object
}
\description{
EmailAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's email address.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

