% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{FieldMetadata}
\alias{FieldMetadata}
\title{FieldMetadata Object}
\usage{
FieldMetadata(primary = NULL, verified = NULL, source = NULL)
}
\arguments{
\item{primary}{True if the field is the primary field; false if the field is a secondary field}

\item{verified}{True if the field is verified; false if the field is unverified}

\item{source}{The source of the field}
}
\value{
FieldMetadata object
}
\description{
FieldMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about a field.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{Gender}}, \code{\link{GetPeopleResponse}},
  \code{\link{ImClient}}, \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

