% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Gender}
\alias{Gender}
\title{Gender Object}
\usage{
Gender(formattedValue = NULL, metadata = NULL, value = NULL)
}
\arguments{
\item{formattedValue}{The read-only value of the gender translated and formatted in the viewer's}

\item{metadata}{Metadata about the gender}

\item{value}{The gender for the person}
}
\value{
Gender object
}
\description{
Gender Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's gender.
}
