% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{ListConnectionsResponse}
\alias{ListConnectionsResponse}
\title{ListConnectionsResponse Object}
\usage{
ListConnectionsResponse(connections = NULL, nextPageToken = NULL,
  nextSyncToken = NULL)
}
\arguments{
\item{connections}{The list of people that the requestor is connected to}

\item{nextPageToken}{The token that can be used to retrieve the next page of results}

\item{nextSyncToken}{The token that can be used to retrieve changes since the last request}
}
\value{
ListConnectionsResponse object
}
\description{
ListConnectionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}}, \code{\link{Locale}},
  \code{\link{Membership}}, \code{\link{Name}},
  \code{\link{Nickname}}, \code{\link{Occupation}},
  \code{\link{Organization}}, \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

