% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Locale}
\alias{Locale}
\title{Locale Object}
\usage{
Locale(metadata = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the locale}

\item{value}{The well-formed [IETF BCP 47](https://tools}
}
\value{
Locale object
}
\description{
Locale Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's locale preference.
}
