% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Membership}
\alias{Membership}
\title{Membership Object}
\usage{
Membership(contactGroupMembership = NULL, domainMembership = NULL,
  metadata = NULL)
}
\arguments{
\item{contactGroupMembership}{The contact group membership}

\item{domainMembership}{The domain membership}

\item{metadata}{Metadata about the membership}
}
\value{
Membership object
}
\description{
Membership Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's read-only membership in a group.
}
