% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Name}
\alias{Name}
\title{Name Object}
\usage{
Name(metadata = NULL, displayName = NULL, familyName = NULL,
  givenName = NULL, middleName = NULL, honorificPrefix = NULL,
  honorificSuffix = NULL, phoneticFamilyName = NULL,
  phoneticGivenName = NULL, phoneticMiddleName = NULL,
  phoneticHonorificPrefix = NULL, phoneticHonorificSuffix = NULL)
}
\arguments{
\item{metadata}{Metadata about the name}

\item{displayName}{The display name formatted according to the locale specified by the viewer's account or the Accept-Language HTTP header}

\item{familyName}{The family name}

\item{givenName}{The given name}

\item{middleName}{The middle name(s)}

\item{honorificPrefix}{The honorific prefixes, such as `Mrs}

\item{honorificSuffix}{The honorific suffixes, such as `Jr}

\item{phoneticFamilyName}{The family name spelled as it sounds}

\item{phoneticGivenName}{The given name spelled as it sounds}

\item{phoneticMiddleName}{The middle name(s) spelled as they sound}

\item{phoneticHonorificPrefix}{The honorific prefixes spelled as they sound}

\item{phoneticHonorificSuffix}{The honorific suffixes spelled as they sound}
}
\value{
Name object
}
\description{
Name Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's name. If the name is a mononym, the family name is empty.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Nickname}}, \code{\link{Occupation}},
  \code{\link{Organization}}, \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

