% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Organization}
\alias{Organization}
\title{Organization Object}
\usage{
Organization(metadata = NULL, type = NULL, formattedType = NULL,
  startDate = NULL, endDate = NULL, current = NULL, name = NULL,
  phoneticName = NULL, department = NULL, title = NULL,
  jobDescription = NULL, symbol = NULL, domain = NULL, location = NULL)
}
\arguments{
\item{metadata}{Metadata about the organization}

\item{type}{The type of the organization}

\item{formattedType}{The read-only type of the organization translated and formatted in the viewer's account locale or the `Accept-Language` HTTP header locale}

\item{startDate}{The start date when the person joined the organization}

\item{endDate}{The end date when the person left the organization}

\item{current}{True if the organization is the person's current organization; false if the organization is a past organization}

\item{name}{The name of the organization}

\item{phoneticName}{The phonetic name of the organization}

\item{department}{The person's department at the organization}

\item{title}{The person's job title at the organization}

\item{jobDescription}{The person's job description at the organization}

\item{symbol}{The symbol associated with the organization; for example, a stock ticker symbol, abbreviation, or acronym}

\item{domain}{The domain name associated with the organization; for example, `google}

\item{location}{The location of the organization office the person works at}
}
\value{
Organization object
}
\description{
Organization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's past or current organization. Overlapping date ranges are permitted.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

