% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Person}
\alias{Person}
\title{Google People API Objects 
The Google People API service gives access to information about profiles and contacts.}
\usage{
Person(resourceName = NULL, etag = NULL, metadata = NULL,
  locales = NULL, names = NULL, nicknames = NULL, coverPhotos = NULL,
  photos = NULL, genders = NULL, ageRange = NULL, birthdays = NULL,
  events = NULL, addresses = NULL, residences = NULL,
  emailAddresses = NULL, phoneNumbers = NULL, imClients = NULL,
  taglines = NULL, biographies = NULL, urls = NULL,
  organizations = NULL, occupations = NULL, interests = NULL,
  skills = NULL, braggingRights = NULL, relations = NULL,
  relationshipInterests = NULL, relationshipStatuses = NULL,
  memberships = NULL)
}
\arguments{
\item{resourceName}{The resource name for the person, assigned by the server}

\item{etag}{The [HTTP entity tag](https://en}

\item{metadata}{Metadata about the person}

\item{locales}{The person's locale preferences}

\item{names}{The person's names}

\item{nicknames}{The person's nicknames}

\item{coverPhotos}{The person's cover photos}

\item{photos}{The person's photos}

\item{genders}{The person's genders}

\item{ageRange}{The person's age range}

\item{birthdays}{The person's birthdays}

\item{events}{The person's events}

\item{addresses}{The person's street addresses}

\item{residences}{The person's residences}

\item{emailAddresses}{The person's email addresses}

\item{phoneNumbers}{The person's phone numbers}

\item{imClients}{The person's instant messaging clients}

\item{taglines}{The person's taglines}

\item{biographies}{The person's biographies}

\item{urls}{The person's associated URLs}

\item{organizations}{The person's past or current organizations}

\item{occupations}{The person's occupations}

\item{interests}{The person's interests}

\item{skills}{The person's skills}

\item{braggingRights}{The person's bragging rights}

\item{relations}{The person's relations}

\item{relationshipInterests}{The kind of relationship the person is looking for}

\item{relationshipStatuses}{The person's relationship statuses}

\item{memberships}{The person's group memberships}
}
\value{
Person object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-04 00:02:09
filename: /Users/mark/dev/R/autoGoogleAPI/googlepeoplev1.auto/R/people_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Person Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a person merged from various data sources such as the authenticated user's contacts and profile data. Fields other than IDs, metadata, and group memberships are user-edited. Most fields can have multiple items. The items in a field have no guaranteed order, but each non-empty field is guaranteed to have exactly one field with `metadata.primary` set to true.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{PhoneNumber}},
  \code{\link{Photo}}, \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

