% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{PersonMetadata}
\alias{PersonMetadata}
\title{PersonMetadata Object}
\usage{
PersonMetadata(sources = NULL, previousResourceNames = NULL,
  deleted = NULL, objectType = NULL)
}
\arguments{
\item{sources}{The sources of data for the person}

\item{previousResourceNames}{Any former resource names this person has had}

\item{deleted}{True if the person resource has been deleted}

\item{objectType}{The type of the person object}
}
\value{
PersonMetadata object
}
\description{
PersonMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about a person.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

