% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{PersonResponse}
\alias{PersonResponse}
\title{PersonResponse Object}
\usage{
PersonResponse(person = NULL, httpStatusCode = NULL,
  requestedResourceName = NULL)
}
\arguments{
\item{person}{The person}

\item{httpStatusCode}{[HTTP 1}

\item{requestedResourceName}{The original requested resource name}
}
\value{
PersonResponse object
}
\description{
PersonResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for a single person
}
