% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{PhoneNumber}
\alias{PhoneNumber}
\title{PhoneNumber Object}
\usage{
PhoneNumber(metadata = NULL, value = NULL, canonicalForm = NULL,
  type = NULL, formattedType = NULL)
}
\arguments{
\item{metadata}{Metadata about the phone number}

\item{value}{The phone number}

\item{canonicalForm}{The read-only canonicalized [ITU-T E}

\item{type}{The type of the phone number}

\item{formattedType}{The read-only type of the phone number translated and formatted in the viewer's account locale or the the `Accept-Language` HTTP header locale}
}
\value{
PhoneNumber object
}
\description{
PhoneNumber Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's phone number.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{Photo}}, \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Source}}, \code{\link{Tagline}},
  \code{\link{Url}}
}

