% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Photo}
\alias{Photo}
\title{Photo Object}
\usage{
Photo(metadata = NULL, url = NULL)
}
\arguments{
\item{metadata}{Metadata about the photo}

\item{url}{The URL of the photo}
}
\value{
Photo object
}
\description{
Photo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's read-only photo. A picture shown next to the person's name tohelp others recognize the person.
}
