% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Relation}
\alias{Relation}
\title{Relation Object}
\usage{
Relation(metadata = NULL, type = NULL, person = NULL,
  formattedType = NULL)
}
\arguments{
\item{metadata}{Metadata about the relation}

\item{type}{The person's relation to the other person}

\item{person}{The name of the other person this relation refers to}

\item{formattedType}{The type of the relation translated and formatted in the viewer's account}
}
\value{
Relation object
}
\description{
Relation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's relation to another person.
}
