% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{RelationshipInterest}
\alias{RelationshipInterest}
\title{RelationshipInterest Object}
\usage{
RelationshipInterest(metadata = NULL, value = NULL, formattedValue = NULL)
}
\arguments{
\item{metadata}{Metadata about the relationship interest}

\item{value}{The kind of relationship the person is looking for}

\item{formattedValue}{The value of the relationship interest translated and formatted in the}
}
\value{
RelationshipInterest object
}
\description{
RelationshipInterest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's read-only relationship interest .
}
