% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{RelationshipStatus}
\alias{RelationshipStatus}
\title{RelationshipStatus Object}
\usage{
RelationshipStatus(metadata = NULL, value = NULL, formattedValue = NULL)
}
\arguments{
\item{metadata}{Metadata about the relationship status}

\item{value}{The relationship status}

\item{formattedValue}{The read-only value of the relationship status translated and formatted in the viewer's account locale or the `Accept-Language` HTTP header locale}
}
\value{
RelationshipStatus object
}
\description{
RelationshipStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's relationship status.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{Relation}}, \code{\link{Residence}},
  \code{\link{Skill}}, \code{\link{Source}},
  \code{\link{Tagline}}, \code{\link{Url}}
}

