% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Residence}
\alias{Residence}
\title{Residence Object}
\usage{
Residence(metadata = NULL, current = NULL, value = NULL)
}
\arguments{
\item{metadata}{Metadata about the residence}

\item{current}{True if the residence is the person's current residence;}

\item{value}{The address of the residence}
}
\value{
Residence object
}
\description{
Residence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A person's past or current residence.
}
