% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(type = NULL, id = NULL)
}
\arguments{
\item{type}{The source type}

\item{id}{A unique identifier within the source type generated by the server}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The source of a field.
}
\seealso{
Other res: \code{\link{Address}}, \code{\link{Biography}},
  \code{\link{Birthday}}, \code{\link{BraggingRights}},
  \code{\link{ContactGroupMembership}},
  \code{\link{CoverPhoto}}, \code{\link{Date}},
  \code{\link{DomainMembership}},
  \code{\link{EmailAddress}}, \code{\link{Event}},
  \code{\link{FieldMetadata}}, \code{\link{Gender}},
  \code{\link{GetPeopleResponse}}, \code{\link{ImClient}},
  \code{\link{Interest}},
  \code{\link{ListConnectionsResponse}},
  \code{\link{Locale}}, \code{\link{Membership}},
  \code{\link{Name}}, \code{\link{Nickname}},
  \code{\link{Occupation}}, \code{\link{Organization}},
  \code{\link{PersonMetadata}},
  \code{\link{PersonResponse}}, \code{\link{Person}},
  \code{\link{PhoneNumber}}, \code{\link{Photo}},
  \code{\link{RelationshipInterest}},
  \code{\link{RelationshipStatus}}, \code{\link{Relation}},
  \code{\link{Residence}}, \code{\link{Skill}},
  \code{\link{Tagline}}, \code{\link{Url}}
}

